requirements.txt
Flask==1.0.2
Flask_Bootstrap4==4.0.2
opencv_python==4.0.0.21
Flask_SQLAlchemy==2.3.2
Jinja2==2.10
Werkzeug==0.14.1
matplotlib==3.0.2
numpy==1.15.1
Hamlish_Jinja==0.3.3
Pillow==5.4.1
flask_bootstrap==3.3.7.1
psycopg2